backend/requirements.txt
# Deep Researcher Agent - Backend Dependencies

# FastAPI and web server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Document processing
PyPDF2==3.0.1
pdfplumber==0.10.3
python-docx==1.1.0
pypandoc==1.11
pytesseract==0.3.10
Pillow==10.1.0

# Vector store and embeddings
faiss-cpu==1.12.0
sentence-transformers==2.2.2
numpy==1.24.3

# LLM integration
openai==1.3.7
transformers==4.35.2
torch==2.1.1

# Report generation
reportlab==4.0.7
markdown==3.5.1
jinja2==3.1.2

# Utilities
python-dotenv==1.0.0
pydantic==2.5.0
httpx==0.25.2
aiofiles==23.2.1

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
